Implemented the UserInteraction class to manage user operations
- This class serves as the main entry point for both librarians and customers interacting with the library system, providing a console-based interface.

- Implemented a main menu that allows users to either:
log in as a librarian or customer, or exit the system. User authentication for librarians is based on a password, while customers are verified using their customer ID.

- Created a dedicated Librarian Menu where authenticated librarians can:
Add or remove books and customers.
Add, return, and renew loans.
Display all library data including books, customers, and loans.
Search for specific books or customer information.
Search for all books loaned to a particular customer.

All data-related operations are handled through calls to the LibrarianInterface class, which interacts with the underlying Database.

- Implemented a CustomerMenu for verified customers where they can:
Return a book.
Search for a book in the library catalog.
View all books currently on loan under their customer ID.

These actions are performed through the CustomerInterface class, which validates user operations based on customer ID.

- Ensured robust input validation for both the main menu and submenus by incorporating checks for valid numerical inputs using the 'GetUserChoice' Method.

The system handles incorrect inputs, prompting users to enter valid options. For librarian and customer logins, appropriate error messages are displayed for incorrect password or invalid customer ID entries, and the user is prompted to try again.

- Utilised DateTime.ParseExact to ensure that dates are entered in the correct format (dd/mm/yyyy), with proper handling of culture-specific date formats using CultureInfo.InvariantCulture.

- Separated functionalities into logical submenus:
'AddRemoveDataMenu' for adding or removing books, customers, and loans.
'DisplayDataMenu' for displaying or searching for books, customers, and loans.